+++ Team 4 - Clothing Store/ClothingStoreData/Delegate/RetrieveMembersDataDelegate.cs
@@ -23,12 +23,12 @@ namespace ClothingStoreData.Delegate
                 members.Add(new Member(
                     reader.GetInt32("MemberID"),
                     reader.GetString("Email"),
-               reader.GetString("FirstName"),
-               reader.GetString("LastName"),
-               reader.GetString("Phone"),
-               reader.GetString("BillingAddress"),
-               reader.GetFloat("Points"),
-               reader.GetString("Status")
+                    reader.GetString("FirstName"),
+                    reader.GetString("LastName"),
+                    reader.GetString("Phone"),
+                    reader.GetString("BillingAddress"),
+                    reader.GetInt32("Points"),
+                    reader.GetString("Status")
                     ));
             }
